De-duplicate CollectGold
