+++ config/symbols.stmad.txt
@@ -21,6 +21,8 @@ g_testCollPrizeTable = 0x80180968;
 g_testCollRandTable = 0x801809A8;
 g_eDamageDisplayClut = 0x801809C4;
 g_bigRedFireballAnim = 0x80180E40;
+g_goldCollectTexts = 0x80180D60;
+c_GoldPrizes = 0x80180D88;
 g_eUnk14SpawnRots = 0x80180FAC;
 unk14_yVel = 0x80180FE4;
 unk14_startFrame = 0x80180FFC;
