+++ config/symbols.us.stcen.txt
@@ -19,6 +19,7 @@ g_testCollPrizeTable = 0x80180B38;
 g_testCollRandTable = 0x80180B78;
 g_eDamageDisplayClut = 0x80180B94;
 g_eRedDoorUV = 0x80180BA4;
+g_goldCollectTexts = 0x80180F30;
 c_GoldPrizes = 0x80180F58;
 c_HeartPrizes = 0x80180FE8;
 g_bigRedFireballAnim = 0x80181014;
