+++ config/symbols.us.stdre.txt
@@ -26,6 +26,7 @@ g_testCollPrizeTable = 0x80180CFC;
 g_testCollRandTable = 0x80180D3C;
 g_eDamageDisplayClut = 0x80180D58;
 g_eRedDoorUV = 0x80180D68;
+g_goldCollectTexts = 0x801810F4;
 c_GoldPrizes = 0x8018111C;
 c_HeartPrizes = 0x801811AC;
 g_bigRedFireballAnim = 0x801811D8;
