+++ config/symbols.us.stnp3.txt
@@ -26,6 +26,7 @@ g_testCollPrizeTable = 0x801819CC;
 g_testCollRandTable = 0x80181A0C;
 g_eDamageDisplayClut = 0x80181A28;
 g_eRedDoorUV = 0x80181A38;
+g_goldCollectTexts = 0x80181DC4;
 c_GoldPrizes = 0x80181DEC;
 c_HeartPrizes = 0x80181E7C;
 g_bigRedFireballAnim = 0x80181EA8;
