+++ src/st/dre/dre.h
@@ -119,8 +119,6 @@ extern u8 D_80181338[];
 extern u16 D_801811A4[];
 extern u16 D_801810B0[];
 extern u16 D_801810E0[];
-extern u8* D_801810F4[];
-extern s32 c_GoldPrizes[];
 
 // *** EntitySoulStealOrb properties START ***
