+++ src/st/no3/no3.h
@@ -183,7 +183,6 @@ extern s16 D_801820C4[];
 extern s16 D_801832D8[];
 extern u8 D_801832E8[];
 extern u16 D_801823F4[];
-extern u8* D_80182438[];
 extern u16 D_80182424[];
 extern u8* D_80182488[];
 extern u16 D_801824E8[];
@@ -194,7 +193,6 @@ extern u16 D_80182628[];
 extern s32 D_80182638[];
 extern u8 D_8018267C[];
 extern ObjInit2 D_80182764[];
-extern s32 c_GoldPrizes[];
 extern u8 D_8018383C[];
 extern s32 D_801838F4[];
 extern s16 D_80183858[];
