+++ src/st/wrp/e_collect.c
@@ -3,11 +3,9 @@
 extern LayoutEntity* D_80180310[];
 extern u16 D_80180DC4[];
 extern u16 D_80180DF4[];
-extern u8* D_80180E08[];
 extern s8 c_HeartPrizes[];
 extern u8* D_80180E58[];
 extern u16 D_80180EB8[];
-extern s32 c_GoldPrizes[];
 extern u16 g_ItemIconSlots[];
 extern u16 UNK_Invincibility0[];
 
@@ -74,27 +72,7 @@ void func_8018CB34(u16 arg0) {
 
 #include "../collect_heart.h"
 
-void CollectGold(u16 goldSize) {
-    s32 *gold, *unk;
-    u16 goldSizeIndex;
-
-    g_api.PlaySfx(NA_SE_PL_COLLECT_GOLD);
-    gold = &g_Status.gold;
-    goldSizeIndex = goldSize - 2;
-    *gold += c_GoldPrizes[goldSizeIndex];
-    if (*gold > MAX_GOLD) {
-        *gold = MAX_GOLD;
-    }
-
-    unk = &g_unkGraphicsStruct.BottomCornerTextTimer;
-    if (*unk) {
-        g_api.FreePrimitives(g_unkGraphicsStruct.BottomCornerTextPrims);
-        *unk = 0;
-    }
-
-    BottomCornerText(D_80180E08[goldSizeIndex], 1);
-    DestroyEntity(g_CurrentEntity);
-}
+#include "../collect_gold.h"
 
 void CollectSubweapon(u16 subWeaponIdx) {
     Entity* player = &PLAYER;
